Python Project for Data Science Final Assignment
PLEASE NOTE : The URL provided for question 4 and Question 6 (GameStop) Stock data does not open the page correctly and hence the webscraping does not work correctly and hence result is not correct and the Graph is also skewed. However i have appended the code with the result so please check the code.